backend/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
python-multipart
ptyprocess; sys_platform == 'linux'
pywinpty; sys_platform == 'win32'
websockets
jinja2
python-dotenv
httpx
pydantic-settings
openai
pycryptodome>=3.19.0
